04-analytics-engineering/pyproject.toml
[project]
name = "04-analytics-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-core>=2.30.3",
    "google-auth>=2.49.2",
    "google-cloud-storage>=3.10.1",
    "google-cloud-bigquery>=3.29.1",
    "pyarrow>=23.0.1",
    "requests>=2.33.1",
]
